src/app/notifications/notifications.service.spec.ts
/**
 * FusionSuite - Frontend
 * Copyright (C) 2022 FusionSuite
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

import { TestBed } from '@angular/core/testing';

import { NotificationsService } from './notifications.service';

describe('NotificationsService', () => {
  let service: NotificationsService;

  beforeEach(() => {
    TestBed.configureTestingModule({});
    service = TestBed.inject(NotificationsService);
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  it('sends success notifications to subscribers', () => {
    service.subscribe((notification) => {
      expect(notification.message).toEqual('A successful message');
      expect(notification.status).toEqual('success');
    });

    service.success('A successful message');
  });

  it('sends error notifications to subscribers', () => {
    service.subscribe((notification) => {
      expect(notification.message).toEqual('An error message');
      expect(notification.status).toEqual('error');
    });

    service.error('An error message');
  });
});
